requirements.txt
pandas
numpy
matplotlib
sklearn.preprocessing
seaborn
sklearn.linear_model